GradeScreen was done for student: - presentation layer package student_grade - GradePoint and GRadePointDto were changed
Participants Screen for both student and teacher was done

Home Page was done

+++ app/src/main/java/com/example/sdo_project/domain/models/GradePoint.kt
@@ -1,10 +1,8 @@
 package com.example.sdo_project.domain.models
 
-import kotlinx.serialization.SerialName
-
 data class GradePoint(
     val id: Int,
     val name: String,
-    val currentScore: Int,
+    val currentScore: Float,
     val maxScore: Int
 )
